+++ app/(main)/parents/profile/layout.tsx
@@ -6,8 +6,9 @@ import {
   SidebarInset,
   SidebarProvider
 } from "@/components/ui/sidebar"
+import { PaperProvider } from "@/context/PaperContext"
 import ParentsNavbar from "@/feature/parents/components/profile/common/Navbar"
-import StudentsNavbar from "@/feature/students/components/dashboard/common/Navbar"
+import StudentsNavbar from "@/feature/students/home/component/Navbar"
 import { BookOpen, Home, ReceiptText, Settings, UserPlus } from "lucide-react"
 
 const projects = [
@@ -41,14 +42,16 @@ const projects = [
 const ParentProfileLayout = ({ children }: { children: React.ReactNode }) => {
   return (
     <SidebarProvider>
-      <AppSidebar projects={projects} />
-      <SidebarInset>
-        {/* <ParentsNavbar /> */}
-        <CustomNavbar role={'parent'} />
-        <div className="flex flex-1 flex-col bg-primary-bg-color gap-4 p-4">
-          {children}
-        </div>
-      </SidebarInset>
+      <PaperProvider>
+        <AppSidebar projects={projects} />
+        <SidebarInset>
+          {/* <ParentsNavbar /> */}
+          <CustomNavbar role={'parent'} />
+          <div className="flex flex-1 flex-col bg-primary-bg-color gap-4 p-4">
+            {children}
+          </div>
+        </SidebarInset>
+      </PaperProvider>
     </SidebarProvider>
   )
 }
