+++ app/(main)/student/setup/layout.tsx
@@ -1,6 +1,6 @@
 // import Navbar from "@/feature/students/components/setup/navbar/Navbar";
 
-import Navbar from "@/feature/students/components/setup/navbar/Navbar";
+import Navbar from "@/feature/students/setup/components/navbar/Navbar";
 
 export default function StudentsSetupLayout({ children }: { children: React.ReactNode }) {
     return (
