+++ app/(main)/student/subjects/[subject]/layout.tsx
@@ -10,7 +10,7 @@ import {
 import { studentSideMenu } from "@/constants/Helpers";
 import { PaperProvider } from "@/context/PaperContext";
 import ParentsNavbar from "@/feature/parents/components/profile/common/Navbar";
-import StudentsNavbar from "@/feature/students/components/dashboard/common/Navbar"
+import StudentsNavbar from "@/feature/students/home/component/Navbar"
 
 const ParentProfileLayout = ({ children }: { children: React.ReactNode }) => {
 
@@ -19,7 +19,10 @@ const ParentProfileLayout = ({ children }: { children: React.ReactNode }) => {
             <PaperProvider>
                 <AppSidebar projects={studentSideMenu} />
                 <SidebarInset>
-                    <CustomNavbar role={'student'} />
+                    <div className="sticky top-0 z-50 ">
+                        <CustomNavbar role={'student'} />
+                    </div>
+                    {/* <CustomNavbar role={'student'} /> */}
                     <div className="flex flex-1 flex-col bg-primary-bg-color gap-4 p-4">
                         {children}
                     </div>
