+++ feature/students/setup/components/Board.tsx
