+++ feature/students/subject/components/IconComponent.tsx
